Add files via upload
A Short Interactive story book. Once a user inputs a username the app allows the user to choose from one of two options to progress through this storybook. Once the button is pressed the user is greeted with a different Image for each page along with sound effects to add suspense while reading. Most of the app is built in swift with little to no code used to build the UI.